refactor(acp): use dynamic import for subcommand
