+++ packages/opencode/src/cli/cmd/acp/handler.ts
@@ -1,0 +1,59 @@
+import * as Log from "@opencode-ai/core/util/log"
+import { bootstrap } from "../../bootstrap"
+import { AgentSideConnection, ndJsonStream } from "@agentclientprotocol/sdk"
+import { ACP } from "@/acp/agent"
+import { Server } from "@/server/server"
+import { createOpencodeClient } from "@opencode-ai/sdk/v2"
+import { resolveNetworkOptions } from "../../network"
+import type { AcpArgs } from "./command"
+
+const log = Log.create({ service: "acp-command" })
+
+export async function handler(args: AcpArgs) {
+  process.env.OPENCODE_CLIENT = "acp"
+  await bootstrap(process.cwd(), async () => {
+    const opts = await resolveNetworkOptions(args)
+    const server = await Server.listen(opts)
+
+    const sdk = createOpencodeClient({
+      baseUrl: `http://${server.hostname}:${server.port}`,
+    })
+
+    const input = new WritableStream<Uint8Array>({
+      write(chunk) {
+        return new Promise<void>((resolve, reject) => {
+          process.stdout.write(chunk, (err) => {
+            if (err) {
+              reject(err)
+            } else {
+              resolve()
+            }
+          })
+        })
+      },
+    })
+    const output = new ReadableStream<Uint8Array>({
+      start(controller) {
+        process.stdin.on("data", (chunk: Buffer) => {
+          controller.enqueue(new Uint8Array(chunk))
+        })
+        process.stdin.on("end", () => controller.close())
+        process.stdin.on("error", (err) => controller.error(err))
+      },
+    })
+
+    const stream = ndJsonStream(input, output)
+    const agent = await ACP.init({ sdk })
+
+    new AgentSideConnection((conn) => {
+      return agent.create(conn, { sdk })
+    }, stream)
+
+    log.info("setup connection")
+    process.stdin.resume()
+    await new Promise((resolve, reject) => {
+      process.stdin.on("end", resolve)
+      process.stdin.on("error", reject)
+    })
+  })
+}
